Dynamically build audio controls
The <audio> elements in the HTML are used in the JS to create the audio
controls. It will place the controls in the same element as its
coresponding <audio> element.

The play/pause button is fully functional.
The time stamp displays the length of the sample but never changes.
The download button will work on any server but not on a local machine
because of CORS restrictions.

The background of the audio controls changes gradient to display
progress of the song as it plays. There is no way to interact/seek
through the song. I don't think there should be (or at least with this
current method) because it will interfere with the buttons.

A possible solution is to add a full seek bar to the bottom border of
the audio controls.